Client - Refactor Code Smell
Unexpected extra semicolon in app.scss
