+++ client/src/App.scss
@@ -284,7 +284,7 @@ body {
   @media (min-width: $screen-w-xxxlg) {
     padding: 17px;
     font-size: 17px;
-  };
+  }
   color: $secondary-gray;
   font-weight: 500;
   border: 1px solid #b1b1b1ab !important;
