+++ client/src/components/device-detail-widgets/processTable.tsx
